fix: do not magically call WithBaseHref when marshalling
For `EmbeddedEntity`, marshalling would result in only the embedded `Entity` being encoded, due to the inheritance of the `MarshalJSON` method. This removes that magic from the API, requiring users to call `WithBaseHref` before encoding to get the same result.

This also refactors the testing to use the black-box approach and uses sub-tests to group test cases by method.